Replace IsMatrixObj by IsMatrix
IsMatrixObj is not supported in old GAP versions (which this package
claims to be compatible with), and it has subtly changed its meaning
in current GAP master. It is also just not needed in any of the cases
that matter for this package -- so this patch just drops any use of it.